--- src/main/kotlin/sliv/tool/data_structures/landmarks/LineLandmark.kt
@@ -1,3 +1,0 @@
-package sliv.tool.data_structures.landmarks
-
-data class LineLandmark(val uid: Long, val startPosition: Point, val endPosition: Point)
