--- src/main/kotlin/sliv/tool/data_structures/landmarks/PlaneLandmark.kt
@@ -1,3 +1,0 @@
-package sliv.tool.data_structures.landmarks
-
-data class PlaneLandmark(val uid: Long, val points: List<Point>)
