--- src/main/kotlin/sliv/tool/data_structures/landmarks/Point.kt
@@ -1,3 +1,0 @@
-package sliv.tool.data_structures.landmarks
-
-data class Point(val xCoordinate: Double, val yCoordinate: Double)
