--- src/main/kotlin/sliv/tool/data_structures/landmarks/PointLandmark.kt
@@ -1,3 +1,0 @@
-package sliv.tool.data_structures.landmarks
-
-data class PointLandmark(val uid: Long, val position: Point)
