+++ src/main/kotlin/sliv/tool/models/landmarks/Point.kt
@@ -1,0 +1,3 @@
+package sliv.tool.models.landmarks
+
+data class Point(val x: Double, val y: Double)
