+++ src/main/kotlin/sliv/tool/models/landmarks/csv/CSVLineLandmark.kt
@@ -1,0 +1,9 @@
+package sliv.tool.models.landmarks.csv
+
+data class CSVLineLandmark(
+    val uid: Long,
+    val x0: Double,
+    val y0: Double,
+    val x1: Double,
+    val y1: Double
+)
