+++ src/main/kotlin/sliv/tool/models/landmarks/csv/CSVPointLandmark.kt
@@ -1,0 +1,3 @@
+package sliv.tool.models.landmarks.csv
+
+data class CSVPointLandmark(val uid: Long, val x: Double, val y: Double)
