+++ src/main/kotlin/sliv/tool/parsers/Parser.kt
@@ -1,0 +1,5 @@
+package sliv.tool.parsers
+
+interface Parser<T> {
+    fun parse(filePath: String): List<T>
+}
