+++ src/main/kotlin/sliv/tool/parsers/lines/CSVLinesParser.kt
@@ -1,0 +1,6 @@
+package sliv.tool.parsers.lines
+
+import sliv.tool.models.landmarks.csv.CSVLineLandmark
+import sliv.tool.parsers.CSVParser
+
+object CSVLinesParser : CSVParser<CSVLineLandmark>(CSVLineLandmark::class.java)
