--- src/main/kotlin/sliv/tool/parsers/lines_parsers/ILinesParser.kt
@@ -1,7 +1,0 @@
-package sliv.tool.parsers.lines_parsers
-
-import sliv.tool.data_structures.landmarks.LineLandmark
-
-interface ILinesParser {
-    fun parse(filePath: String): List<LineLandmark>
-}
