--- src/main/kotlin/sliv/tool/parsers/planes_parsers/IPlanesParser.kt
@@ -1,7 +1,0 @@
-package sliv.tool.parsers.planes_parsers
-
-import sliv.tool.data_structures.landmarks.PointLandmark
-
-interface IPlanesParser {
-    fun parse(filePath: String): List<PointLandmark>
-}
