--- src/main/kotlin/sliv/tool/parsers/planes_parsers/ImagePlanesParser.kt
@@ -1,9 +1,0 @@
-package sliv.tool.parsers.planes_parsers
-
-import sliv.tool.data_structures.landmarks.PointLandmark
-
-class ImagePlanesParser : IPlanesParser {
-    override fun parse(filePath: String): List<PointLandmark> {
-        TODO()
-    }
-}
