+++ src/main/kotlin/sliv/tool/parsers/points/CSVPointsParser.kt
@@ -1,0 +1,6 @@
+package sliv.tool.parsers.points
+
+import sliv.tool.models.landmarks.csv.CSVPointLandmark
+import sliv.tool.parsers.CSVParser
+
+object CSVPointsParser : CSVParser<CSVPointLandmark>(CSVPointLandmark::class.java)
