--- src/main/kotlin/sliv/tool/parsers/points_parsers/IPointsParser.kt
@@ -1,7 +1,0 @@
-package sliv.tool.parsers.points_parsers
-
-import sliv.tool.data_structures.landmarks.PointLandmark
-
-interface IPointsParser {
-    fun parse(filePath: String): List<PointLandmark>
-}
